refactor: replace FileUpload component with FormImageUpload across multiple forms
