+++ components/elements/form-image-upload/index.ts
@@ -1,0 +1,1 @@
+export { FormImageUpload } from './form-image-upload'
