--- components/elements/image-upload/index.ts
@@ -1,1 +1,0 @@
-export { ImageUpload } from './image-upload'
