+++ components/elements/index.ts
@@ -4,9 +4,9 @@ export { CategoryIcon } from './category-icon'
 export { CheckConsent } from './cookie-consent/check-consent'
 export { CookieConsent } from './cookie-consent/cookie-consent'
 export { ManageCookies } from './cookie-consent/manage-cookies'
+export { FormImageUpload } from './form-image-upload'
 export { HeadingImage } from './heading-image'
 export { IconUpload } from './icon-upload'
-export { ImageUpload } from './image-upload'
 export { IsAdmin, isAdmin } from './is-admin'
 export { MD } from './md'
 export { ImageLightbox } from './modal'
